+++ lib/rolling_switch.dart
@@ -1,4 +1,5 @@
 library rolling_switch;
 
+export 'src/controller/rolling_controller.dart';
 export 'src/info/rolling_info.dart';
 export 'src/rolling_switch_widget.dart';